+++ src/test/java/io/github/astrapisixtynine/pdf/to/text/pdfbox/PdfToTextExtensionsTest.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapisixtynine.pdf.to.text;
+package io.github.astrapisixtynine.pdf.to.text.pdfbox;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -40,7 +40,6 @@ import io.github.astrapi69.file.create.FileFactory;
 import io.github.astrapi69.file.delete.DeleteFileExtensions;
 import io.github.astrapi69.file.search.PathFinder;
 import io.github.astrapisixtynine.pdf.to.text.info.ConversionResult;
-import net.sourceforge.tess4j.TesseractException;
 
 /**
  * Test class for {@link PdfToTextExtensions}
@@ -101,6 +100,7 @@ class PdfToTextExtensionsTest
 		File resultDir = PathFinder.getRelativePath(PathFinder.getSrcTestResourcesDir(),
 			"text-result");
 		File resultTextFile = PdfToTextExtensions.pdfToText(pdfFile, resultDir);
+		assertNotNull(resultTextFile);
 	}
 
 	/**
@@ -143,38 +143,6 @@ class PdfToTextExtensionsTest
 
 		DeleteFileExtensions.delete(conversionResult.getImageFiles());
 		DeleteFileExtensions.delete(conversionResult.getTextFiles());
-	}
-
-	/**
-	 * Tests the method {@link PdfToTextExtensions#extractTextFromImage(File, String, String)}
-	 */
-	@Test
-	void extractTextFromImage() throws TesseractException
-	{
-		String language = "deu";
-		File imageFile = new File("src/test/resources/sample-image.png");
-		String datapath = "/usr/share/tessdata";
-		String result = PdfToTextExtensions.extractTextFromImage(imageFile, datapath, language);
-		assertNotNull(result);
-		assertFalse(result.isEmpty());
-	}
-
-	/**
-	 * Tests if the method correctly detects when Tesseract is installed. This test will pass if
-	 * Tesseract is installed and available in the system's PATH.
-	 */
-	@Test
-	@Disabled("""
-		only for local use:
-		if tesseract is installed
-		""")
-	public void testIsTesseractInstalledWhenInstalled()
-	{
-		// This test assumes that Tesseract is installed on the system.
-		boolean isInstalled = PdfToTextExtensions.isTesseractInstalled();
-
-		// Tesseract should be installed, so the result should be true.
-		assertTrue(isInstalled, "Tesseract should be installed on the system.");
 	}
 
 	/**